Extract the SupportItem from The ApplicationBar
